renaming to postsQueries
